refactored out downloading functionality from buttons to useDownloadMedia hook
